+++ modules/apps/portal-security-sso-google/portal-security-sso-google-impl/src/main/java/com/liferay/portal/security/sso/google/internal/module/settings/definition/GoogleAuthorizationCompanyServiceConfigurationBeanDeclaration.java
@@ -12,23 +12,23 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.openid.internal.module.configuration.definition;
+package com.liferay.portal.security.sso.google.internal.module.settings.definition;
 
 import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
-import com.liferay.portal.security.sso.openid.configuration.OpenIdConfiguration;
+import com.liferay.portal.security.sso.google.configuration.GoogleAuthorizationConfiguration;
 
 import org.osgi.service.component.annotations.Component;
 
 /**
- * @author Mika Koivisto
+ * @author Stian Sigvartsen
  */
 @Component(service = ConfigurationBeanDeclaration.class)
-public class OpenIdCompanyServiceConfigurationBeanDeclaration
+public class GoogleAuthorizationCompanyServiceConfigurationBeanDeclaration
 	implements ConfigurationBeanDeclaration {
 
 	@Override
 	public Class<?> getConfigurationBeanClass() {
-		return OpenIdConfiguration.class;
+		return GoogleAuthorizationConfiguration.class;
 	}
 
 }
