+++ modules/apps/portal-security-sso/portal-security-sso-opensso-impl/src/main/java/com/liferay/portal/security/sso/opensso/internal/module/settings/definition/OpenSSOCompanyServiceConfigurationBeanDeclaration.java
@@ -12,29 +12,23 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.openid.connect.internal.module.configuration.declaration;
+package com.liferay.portal.security.sso.opensso.internal.module.settings.definition;
 
-import com.liferay.portal.kernel.settings.definition.ConfigurationPidMapping;
-import com.liferay.portal.security.sso.openid.connect.configuration.OpenIdConnectConfiguration;
-import com.liferay.portal.security.sso.openid.connect.constants.OpenIdConnectConstants;
+import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
+import com.liferay.portal.security.sso.opensso.configuration.OpenSSOConfiguration;
 
 import org.osgi.service.component.annotations.Component;
 
 /**
- * @author Michael C. Han
+ * @author Mika Koivisto
  */
-@Component(immediate = true, service = ConfigurationPidMapping.class)
-public class OpenIdConnectCompanyServiceConfigurationPidMapping
-	implements ConfigurationPidMapping {
+@Component(service = ConfigurationBeanDeclaration.class)
+public class OpenSSOCompanyServiceConfigurationBeanDeclaration
+	implements ConfigurationBeanDeclaration {
 
 	@Override
 	public Class<?> getConfigurationBeanClass() {
-		return OpenIdConnectConfiguration.class;
-	}
-
-	@Override
-	public String getConfigurationPid() {
-		return OpenIdConnectConstants.SERVICE_NAME;
+		return OpenSSOConfiguration.class;
 	}
 
 }
